break endpoints out into multiple routers
